# frontend/openapi.json
{
 "components": {
  "schemas": {
...
   "AreaIn": {
    "properties": {
     "code": {
      "maxLength": 40,
      "minLength": 1,
      "title": "Code",
      "type": "string"
     },
     "name": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "site_id": {
      "title": "Site Id",
      "type": "integer"
     }
    },
    "required": [
     "site_id",
     "name",
     "code"
    ],
    "title": "AreaIn",
    "type": "object"
   },
   "AreaOut": {
    "properties": {
     "archived_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Archived At"
     },
     "code": {
      "title": "Code",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "site_id": {
      "title": "Site Id",
      "type": "integer"
     },
     "updated_at": {
      "format": "date-time",
      "title": "Updated At",
      "type": "string"
     }
    },
    "required": [
     "id",
     "site_id",
     "name",
     "code",
     "archived_at",
     "updated_at"
    ],
    "title": "AreaOut",
    "type": "object"
   },
   "AreaPatch": {
    "properties": {
     "archived": {
      "anyOf": [
       {
        "type": "boolean"
       },
       {
        "type": "null"
       }
      ],
      "title": "Archived"
     },
     "code": {
      "anyOf": [
       {
        "maxLength": 40,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Code"
     },
     "name": {
      "anyOf": [
       {
        "maxLength": 200,
        "minLength": 1,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     }
    },
    "title": "AreaPatch",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "OrgSettings": {
    "description": "Per-organization configuration. Stored as JSON on organizations.settings; defaults fill gaps.",
    "properties": {
     "alert_channels": {
      "default": [
       "in_app"
      ],
      "items": {
       "enum": [
        "in_app",
        "email",
        "push",
        "whatsapp",
        "sms"
       ],
       "type": "string"
      },
      "title": "Alert Channels",
      "type": "array"
     },
     "alert_thresholds": {
      "$ref": "#/components/schemas/AlertThresholds",
      "default": {
       "long_stop_manager_min": 60,
       "long_stop_supervisor_min": 15,
       "oee_below_pct": 50.0,
       "reject_rate_above_pct": 5.0,
       "unexplained_gap_min": 20
      }
     },
     "allow_negative_stock": {
      "default": false,
      "title": "Allow Negative Stock",
      "type": "boolean"
     },
     "backflush_on_output": {
      "default": false,
      "title": "Backflush On Output",
      "type": "boolean"
     },
     "currency": {
      "default": "INR",
      "title": "Currency",
      "type": "string"
     },
     "data_retention_days": {
      "default": 1095,
      "title": "Data Retention Days",
      "type": "integer"
     },
     "default_language": {
      "default": "en",
      "title": "Default Language",
      "type": "string"
     },
     "digest_time": {
      "default": "07:00",
      "title": "Digest Time",
      "type": "string"
     },
     "low_stock_default_days": {
      "default": 7,
      "title": "Low Stock Default Days",
      "type": "integer"
     },
     "oee_ideal_rate_mode": {
      "default": "product",
      "enum": [
       "machine",
       "product"
      ],
      "title": "Oee Ideal Rate Mode",
      "type": "string"
     },
     "operator_idle_lock_min": {
      "default": 10,
      "maximum": 480.0,
      "minimum": 1.0,
      "title": "Operator Idle Lock Min",
      "type": "integer"
     },
     "planned_changeovers": {
      "default": false,
      "title": "Planned Changeovers",
      "type": "boolean"
     },
     "require_2fa_for_admins": {
      "default": false,
      "title": "Require 2Fa For Admins",
      "type": "boolean"
     },
     "show_operator_analytics": {
      "default": false,
      "title": "Show Operator Analytics",
      "type": "boolean"
     },
     "timezone": {
      "default": "Asia/Kolkata",
      "title": "Timezone",
      "type": "string"
     },
     "unit_system": {
      "default": "metric",
      "enum": [
       "metric",
       "imperial"
      ],
      "title": "Unit System",
      "type": "string"
     },
     "week_start_day": {
      "default": 0,
      "maximum": 6.0,
      "minimum": 0.0,
      "title": "Week Start Day",
      "type": "integer"
     },
     "weekly_summary_day": {
      "default": 0,
      "maximum": 6.0,
      "minimum": 0.0,
      "title": "Weekly Summary Day",
      "type": "integer"
     },
     "working_days": {
      "default": [
       0,
       1,
       2,
       3,
       4,
       5
      ],
      "items": {
       "type": "integer"
      },
      "title": "Working Days",
      "type": "array"
     }
    },
    "title": "OrgSettings",
    "type": "object"
   },
   "Page_AuditOut_": {
    "properties": {
     "items": {
      "items": {
       "$ref": "#/components/schemas/AuditOut"
      },
      "title": "Items",
      "type": "array"
     },
     "page": {
      "title": "Page",
      "type": "integer"
     },
     "page_size": {
      "title": "Page Size",
      "type": "integer"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "items",
     "total",
     "page",
     "page_size"
    ],
    "title": "Page[AuditOut]",
    "type": "object"
   },
...
   "Role": {
    "enum": [
     "operator",
     "supervisor",
     "technician",
     "inspector",
     "storekeeper",
     "manager",
     "admin"
    ],
    "title": "Role",
    "type": "string"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "FloorPulse API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
  "/api/v1/areas": {
   "get": {
    "operationId": "list_areas_api_v1_areas_get",
    "parameters": [
     {
      "in": "query",
      "name": "site_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Site Id"
      }
     },
     {
      "in": "query",
      "name": "include_archived",
      "required": false,
      "schema": {
       "default": false,
       "title": "Include Archived",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/AreaOut"
         },
         "title": "Response List Areas Api V1 Areas Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Areas",
    "tags": [
     "admin"
    ]
   },
   "post": {
    "operationId": "create_area_api_v1_areas_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AreaIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AreaOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Create Area",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/areas/{area_id}": {
   "patch": {
    "operationId": "patch_area_api_v1_areas__area_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "area_id",
      "required": true,
      "schema": {
       "title": "Area Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "if-match",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AreaPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AreaOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Patch Area",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/audit": {
   "get": {
    "operationId": "list_audit_api_v1_audit_get",
    "parameters": [
     {
      "in": "query",
      "name": "entity_type",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Type"
      }
     },
     {
      "in": "query",
      "name": "entity_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Entity Id"
      }
     },
     {
      "in": "query",
      "name": "actor_user_id",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ],
       "title": "Actor User Id"
      }
     },
     {
      "in": "query",
      "name": "action",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Action"
      }
     },
     {
      "in": "query",
      "name": "since",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Since"
      }
     },
     {
      "in": "query",
      "name": "until",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "format": "date-time",
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Until"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 500,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_AuditOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Audit",
    "tags": [
     "admin"
    ]
   }
  },
...
  "/api/v1/org/settings": {
   "get": {
    "operationId": "get_org_settings_api_v1_org_settings_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrgSettings"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Org Settings",
    "tags": [
     "admin"
    ]
   },
   "put": {
    "operationId": "put_org_settings_api_v1_org_settings_put",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OrgSettings"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrgSettings"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Put Org Settings",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/sites": {
   "get": {
    "operationId": "list_sites_api_v1_sites_get",
    "parameters": [
     {
      "in": "query",
      "name": "include_archived",
      "required": false,
      "schema": {
       "default": false,
       "title": "Include Archived",
       "type": "boolean"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SiteOut"
         },
         "title": "Response List Sites Api V1 Sites Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Sites",
    "tags": [
     "admin"
    ]
   },
   "post": {
    "operationId": "create_site_api_v1_sites_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SiteIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SiteOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Create Site",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/sites/{site_id}": {
   "get": {
    "operationId": "get_site_api_v1_sites__site_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SiteOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get Site",
    "tags": [
     "admin"
    ]
   },
   "patch": {
    "operationId": "patch_site_api_v1_sites__site_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "site_id",
      "required": true,
      "schema": {
       "title": "Site Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "if-match",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SitePatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SiteOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Patch Site",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/users": {
   "get": {
    "operationId": "list_users_api_v1_users_get",
    "parameters": [
     {
      "in": "query",
      "name": "q",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Q"
      }
     },
     {
      "in": "query",
      "name": "role",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "$ref": "#/components/schemas/Role"
        },
        {
         "type": "null"
        }
       ],
       "title": "Role"
      }
     },
     {
      "in": "query",
      "name": "active",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "boolean"
        },
        {
         "type": "null"
        }
       ],
       "title": "Active"
      }
     },
     {
      "in": "query",
      "name": "page",
      "required": false,
      "schema": {
       "default": 1,
       "minimum": 1,
       "title": "Page",
       "type": "integer"
      }
     },
     {
      "in": "query",
      "name": "page_size",
      "required": false,
      "schema": {
       "default": 50,
       "maximum": 500,
       "minimum": 1,
       "title": "Page Size",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Page_UserOut_"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "List Users",
    "tags": [
     "admin"
    ]
   },
   "post": {
    "operationId": "create_user_api_v1_users_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Create User",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/users/{user_id}": {
   "get": {
    "operationId": "get_user_api_v1_users__user_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "title": "User Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Get User",
    "tags": [
     "admin"
    ]
   },
   "patch": {
    "operationId": "patch_user_api_v1_users__user_id__patch",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "title": "User Id",
       "type": "integer"
      }
     },
     {
      "in": "header",
      "name": "if-match",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "If-Match"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserPatch"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Patch User",
    "tags": [
     "admin"
    ]
   }
  },
  "/api/v1/users/{user_id}/logout-all": {
   "post": {
    "operationId": "logout_user_everywhere_api_v1_users__user_id__logout_all_post",
    "parameters": [
     {
      "in": "path",
      "name": "user_id",
      "required": true,
      "schema": {
       "title": "User Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Logout User Everywhere",
    "tags": [
     "admin"
    ]
   }
  },
...
 }
}